+++ src/main/java/kr/sesacjava/swimtutor/record/repository/ExerciseRecordRepository.java
@@ -1,0 +1,8 @@
+package kr.sesacjava.swimtutor.record.repository;
+
+import kr.sesacjava.swimtutor.record.entity.ExerciseRecord;
+import kr.sesacjava.swimtutor.record.entity.ExerciseRecordId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ExerciseRecordRepository extends JpaRepository<ExerciseRecord, ExerciseRecordId> {
+}
