--- src/main/java/kr/sesacjava/swimtutor/repository/UserExerciseRecordRepository.java
@@ -1,8 +1,0 @@
-package kr.sesacjava.swimtutor.repository;
-
-import kr.sesacjava.swimtutor.domain.UserExerciseRecord;
-import kr.sesacjava.swimtutor.domain.UserExerciseRecordId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserExerciseRecordRepository extends JpaRepository<UserExerciseRecord, UserExerciseRecordId> {
-}
